Fix Delete and Reset Issues
The Delete Icon was not showing up on the Edit Timer Page for some sections
since the SetTimer method needed to be set before the timer section. The Reset
behavior for the view timer page was changed so it will leave the page if the
timer has not previously been saved.